+++ astropy/time/core.py
@@ -1706,9 +1706,11 @@ class TimeDelta(Time):
 
     Parameters
     ----------
-    val : sequence, ndarray, number, or `~astropy.time.TimeDelta` object
-        Value(s) to initialize the time difference(s).
-    val2 : numpy ndarray, list, str, or number; optional
+    val : sequence, ndarray, number, `~astropy.units.Quantity` or `~astropy.time.TimeDelta` object
+        Value(s) to initialize the time difference(s). Any quantities will
+        be converted appropriately (with care taken to avoid rounding
+        errors for regular time units).
+    val2 : sequence, ndarray, number, or `~astropy.units.Quantity`; optional
         Additional values, as needed to preserve precision.
     format : str, optional
         Format of input value(s)
@@ -1737,14 +1739,6 @@ class TimeDelta(Time):
                 self._set_scale(scale)
         else:
             if format is None:
-                try:
-                    val = val.to(u.day)
-                    if val2 is not None:
-                        val2 = val2.to(u.day)
-                except Exception:
-                    raise ValueError('Only Quantities with Time units can '
-                                     'be used to initiate {0} instances .'
-                                     .format(self.__class__.__name__))
                 format = 'jd'
 
             self._init_from_vals(val, val2, format, scale, copy)