+++ CHANGES.rst
@@ -1183,6 +1183,8 @@ astropy.tests
 astropy.time
 ^^^^^^^^^^^^
 
+- Avoid rounding errors when converting ``Quantity`` to ``TimeDelta``. [#7625]
+
 astropy.units
 ^^^^^^^^^^^^^